added cases level for ticket types
