+++ data-access/src/app/application-services/cases/violation-ticket/index.ts
@@ -1,0 +1,14 @@
+import { AppContext } from '../../../init/app-context-builder';
+import { ViolationTicketV1Api, ViolationTicketV1ApiImpl } from './v1';
+
+export interface ViolationTicketApi { 
+  v1: ViolationTicketV1Api;
+}
+
+export class ViolationTicketApiImpl implements ViolationTicketApi {
+  v1: ViolationTicketV1Api;
+
+  constructor(context: AppContext) {
+    this.v1 = new ViolationTicketV1ApiImpl(context);
+  }
+}
