+++ data-access/src/app/external-dependencies/datastore.ts
@@ -3,6 +3,6 @@ export { User as UserData, UserModel } from "../../infrastructure-services-impl/
 export { Role as RoleData, RoleModel, CommunityPermissions } from "../../infrastructure-services-impl/datastore/mongodb/models/role";
 export { Property as PropertyData, PropertyModel } from "../../infrastructure-services-impl/datastore/mongodb/models/property";
 export { Service as ServiceData, ServiceModel } from "../../infrastructure-services-impl/datastore/mongodb/models/service";
-export { ServiceTicket as ServiceTicketData, ServiceTicketModel } from "../../infrastructure-services-impl/datastore/mongodb/models/service-ticket";
+export { ServiceTicket as ServiceTicketData, ServiceTicketModel } from "../../infrastructure-services-impl/datastore/mongodb/models/cases/service-ticket";
 export { Member as MemberData, MemberModel } from "../../infrastructure-services-impl/datastore/mongodb/models/member";
-export { ViolationTicket as ViolationTicketData, ViolationTicketModel } from "../../infrastructure-services-impl/datastore/mongodb/models/violation-ticket";
+export { ViolationTicket as ViolationTicketData, ViolationTicketModel } from "../../infrastructure-services-impl/datastore/mongodb/models/cases/violation-ticket";
