+++ data-access/src/app/init/application-services-builder.ts
@@ -3,12 +3,11 @@ import {
   CommunityApi,
   MemberApi,
   PropertyApi,
-  ServiceTicketApi,
   PaymentApi,
-  ViolationTicketApi,
   ServiceApi,
   RoleApi,
   UserApi,
+  CasesApi,
 } from "../application-services";
 import { AppContext } from "./app-context-builder";
 import { CommunityApiImpl } from "../application-services/community";
@@ -16,17 +15,15 @@ import { MemberApiImpl } from "../application-services/member";
 import { PropertyApiImpl } from "../application-services/property";
 import { RoleApiImpl } from "../application-services/role";
 import { ServiceApiImpl } from "../application-services/service";
-import { ServiceTicketApiImpl } from "../application-services/service-ticket";
 import { UserApiImpl } from "../application-services/user";
-import { ViolationTicketApiImpl } from "../application-services/violation-ticket";
 import { PaymentApiImpl } from "../application-services/payment";
+import { CasesApiImpl } from "../application-services/cases";
 ;
 export class ApplicationServicesBuilder implements ApplicationServices {
   community: CommunityApi;
   member: MemberApi;
   property: PropertyApi;
-  serviceTicket: ServiceTicketApi;
-  violationTicket: ViolationTicketApi;
+  cases: CasesApi;
   service: ServiceApi;
   role: RoleApi;
   user: UserApi;
@@ -36,8 +33,7 @@ export class ApplicationServicesBuilder implements ApplicationServices {
     this.community = new CommunityApiImpl(context);
     this.member = new MemberApiImpl(context);
     this.property = new PropertyApiImpl(context);
-    this.serviceTicket = new ServiceTicketApiImpl(context);
-    this.violationTicket = new ViolationTicketApiImpl(context);
+    this.cases = new CasesApiImpl(context);
     this.service = new ServiceApiImpl(context);
     this.role = new RoleApiImpl(context);
     this.user = new UserApiImpl(context);
