--- data-access/src/infrastructure-services-impl/datastore/memorydb/infrastructure/violation-ticket/v1/violation-ticket.memory-uow.ts
@@ -1,11 +1,0 @@
-import { ViolationTicketV1, ViolationTicketV1Props } from "../../../../../../app/domain/contexts/cases/violation-ticket/v1/violation-ticket";
-import { MemoryStore } from "../../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-store";
-import { MemoryUnitOfWork } from "../../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-unit-of-work";
-import { InProcEventBusInstance, NodeEventBusInstance } from "../../../../../../../seedwork/event-bus-seedwork-node";
-import { MemoryViolationTicketV1Repository } from "./violation-ticket.memory-repository";
-
-export const buildMemoryViolationTicketV1UnitOfWork = (
-  violationTicketV1MemoryStore: MemoryStore<ViolationTicketV1Props>,
-) => {
-  return new MemoryUnitOfWork(InProcEventBusInstance, NodeEventBusInstance, ViolationTicketV1, violationTicketV1MemoryStore, MemoryViolationTicketV1Repository);
-};
